Extract kernel.stream() stream merge logic into utils
