+++ lib/core/kernel.js
@@ -17,12 +17,12 @@
 const _ = require('lodash')
 const debug = require('debug')('jellyfish:kernel')
 const Bluebird = require('bluebird')
-const EventEmitter = require('events').EventEmitter
 const jsonSchema = require('./json-schema')
 const errors = require('./errors')
 const CARDS = require('./cards')
 const permissionFilter = require('./permission-filter')
 const formulas = require('./formulas')
+const utils = require('./utils')
 
 const getFilteredQuery = async (kernel, session, schema, options = {}) => {
 	return permissionFilter.getQuery(kernel.backend, session, schema, {
@@ -449,32 +449,9 @@ module.exports = class Kernel {
    */
 	async stream (session, schema, options) {
 		const filteredQuery = await getFilteredQuery(this, session, schema, options)
-		const emitter = new EventEmitter()
-		const streams = {}
-
-		for (const bucket of this.getBucketsForSchema(schema)) {
-			streams[bucket] = await this.backend.stream(bucket, filteredQuery)
-			streams[bucket].on('closed', () => {
-				Reflect.deleteProperty(streams, bucket)
-				if (_.isEmpty(streams)) {
-					emitter.emit('closed')
-				}
-			})
-
-			for (const event of [ 'error', 'data' ]) {
-				streams[bucket].on(event, (data) => {
-					emitter.emit(event, data)
-				})
-			}
-		}
-
-		emitter.close = () => {
-			_.each(streams, (stream) => {
-				stream.close()
-			})
-		}
-
-		return emitter
+		return utils.mergeStreams(this.getBucketsForSchema(schema), (bucket) => {
+			return this.backend.stream(bucket, filteredQuery)
+		})
 	}
 
 	/**
