+++ lib/core/utils.js
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+const _ = require('lodash')
+const EventEmitter = require('events').EventEmitter
 const objectHash = require('object-hash')
 
 /**
@@ -36,4 +38,35 @@ exports.hashObject = (object) => {
 		// This in particular is a HUGE improvement
 		respectType: false
 	})
+}
+
+exports.mergeStreams = async (keys, createFunction) => {
+	const emitter = new EventEmitter()
+	const streams = {}
+
+	for (const key of keys) {
+		streams[key] = await createFunction(key)
+		streams[key].on('closed', () => {
+			Reflect.deleteProperty(streams, key)
+			if (_.isEmpty(streams)) {
+				emitter.emit('closed')
+			}
+		})
+
+		streams[key].on('error', (error) => {
+			emitter.emit('error', error)
+		})
+
+		streams[key].on('data', (data) => {
+			emitter.emit('data', data)
+		})
+	}
+
+	emitter.close = () => {
+		_.each(streams, (stream) => {
+			stream.close()
+		})
+	}
+
+	return emitter
 }
